sysintegral_WEB_SERVER/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
pymysql==1.1.1
apscheduler==3.10.4
